requirements.txt
Flask
gunicorn
easyocr
opencv-python-headless
numpy
Pillow
firebase-admin
os
json
